pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "synaptic"
version = "0.1.0"
description = "An orchestration framework for large language models"
authors = [{ name = "phoenix5971" }]
dependencies = [
	"anthropic",
	"openai",
	"google-generativeai",
	"google-genai",
	"dotenv",
	"vertexai",
	"xai-sdk",
]

[project.optional-dependencies]
mcp = ["mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
